backend/requirements.txt
alembic==1.13.3
amqp==5.2.0
annotated-types==0.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==2.4.1
async-timeout==5.0.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
bcrypt==4.3.0
billiard==4.2.1
blinker==1.8.2
cachelib==0.9.0
celery==5.4.0
cffi==1.17.1
channels==4.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.0
cxxfilt==0.3.0
daphne==4.1.2
decorator==5.1.1
defusedxml==0.7.1
dissect==3.19
Django==5.1.4
django-celery-results==2.5.1
django-cors-headers==4.6.0
django-db==0.0.7
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dncil==1.0.2
dnfile==0.15.1
dnspython==2.7.0
drf-yasg==1.21.8
email-validator==1.0.5
executing==2.1.0
flare-capa==8.0.1
flow.record==3.19
funcy==2.0
fusepy==3.0.1
greenlet==3.1.1
gunicorn==23.0.0
humanize==4.11.0
hyperlink==21.0.0
ida-netnode==3.0
ida-settings==2.1.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
intervaltree==3.1.0
ipython==8.29.0
isbnlib==3.9.10
iso3166==1.0.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
kombu==5.4.2
Mako==1.3.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
msgpack==1.0.8
msgspec==0.19.0
networkx==3.4.2
packaging==24.2
paramiko==3.5.1
parso==0.8.4
pefile==2024.8.26
pexpect==4.9.0
prompt_toolkit==3.0.48
protobuf==5.29.2
psutil==6.1.1
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py-moneyed==0.8.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.10.4
pydantic_core==2.27.2
pyelftools==0.31
Pygments==2.18.0
PyJWT==2.9.0
pymemcache==4.0.0
PyMySQL==1.1.1
PyNaCl==1.5.0
pyOpenSSL==24.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-flirt==0.8.10
python-magic==0.4.27
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
rest-framework-simplejwt==0.0.2
rich==13.9.4
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
schwifty==2020.2.2
service-identity==24.2.0
six==1.16.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
sqlparse==0.5.3
stack-data==0.6.3
structlog==24.4.0
traitlets==5.14.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
vine==5.1.0
viv_utils==0.7.12
vivisect==1.2.1
wcwidth==0.2.13
Werkzeug==2.2.2
xmltodict==0.14.2
zope.interface==7.2
zstandard==0.23.0
yara-python
docutils
